requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
httpx
